+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added: Element#getExtendedData - without table name, to get all extended data rows for that element 
 * Added: Fetch hints to get old metadata
 * Fixed: "Scanner already closed" exception in ScannerStreamUtils
+* Fixed: Query on Extended Data reserved field names (ELEMENT_TYPE, ELEMENT_ID, TABLE_NAME, ROW_ID)
 * Changed: Speed up element delete when no extended data rows are present
 
 # v4.6.3